cybeXtelegram/sourukorekuta/requirements.txt
telethon
psycopg2
python-telegram-bot==8.1.1
beautifulsoup4==4.6.0
requests==2.13.0
datefinder==0.6.1
dateparser==0.6.0
six==1.10.0
urllib3==1.22
python-dateutil==2.6.1
feedparser==5.2.1
gdax==1.0.6
